Bring alphabet index UI from app drawer to home
